requirements.txt
botocore #no-deploy
boto3 #no-deploy
Flask==1.1.1
pynamodb==4.0.0
pytest-mock==1.10.4